fix(api): keep core DTO imports side-effect free
